refactor(async-utils): move fake timer advancement into seperate utility
