--- src/helpers/jestFakeTimersAreEnabled.ts
@@ -1,13 +1,0 @@
-export const jestFakeTimersAreEnabled = () => {
-  /* istanbul ignore else */
-  if (typeof jest !== 'undefined' && jest !== null) {
-    return (
-      // legacy timers
-      jest.isMockFunction(setTimeout) ||
-      // modern timers
-      Object.prototype.hasOwnProperty.call(setTimeout, 'clock')
-    )
-  }
-  // istanbul ignore next
-  return false
-}
